@@ -45,5 +45,5 @@ func DeleteData(ctx context.Context, appPaths *paths.Paths, removeKubernetesCach
 	} else {
 		pathList = append(pathList, filepath.Join(appPaths.Cache, "updater-longhorn.json"))
 	}
-	return deleteUnixLikeData(appPaths, pathList)
+	return deleteUnixLikeData(ctx, appPaths, pathList)
 }
