@@ -17,7 +17,7 @@ func DeleteData(ctx context.Context, appPaths *paths.Paths, removeKubernetesCach
 	if err := autostart.EnsureAutostart(ctx, false); err != nil {
 		logrus.Errorf("Failed to remove autostart configuration: %s", err)
 	}
-	if err := deleteLimaVM(); err != nil && !errors.Is(err, os.ErrNotExist) {
+	if err := deleteLimaVM(ctx); err != nil && !errors.Is(err, os.ErrNotExist) {
 		logrus.Errorf("Failed to delete Lima VM: %s", err)
 	}
 	w := wsl.WSLImpl{}
